+++ docs/Changelog.md
@@ -1,10 +1,22 @@
 # Changelog
 
+## 0.2.15
+  - replace digits and UUID sequences with N and U to decrease label count
+
+## 0.2.14
+  - added support for user defined variables
+  - fix reading for non-standard errors JTL
+
+## 0.2.13
+  - Some more stats have been added to console screen
+  - add `-gui` command-line alias to open JMeter UI for debugging
+  - add support for JMeter distributed mode
+  
 ## 0.2.12
-  - Added http request defaults options.
+  - Added http request defaults options
   - Added support of RPS shaper component
   - Remove conflicting JAR libraries during JMeter installation procedure
-  - add GUI window for status screen on Windows
+  - Fixed bug when resource files were not properly copied to artifacts directory
 
 ## 0.2.11
   - Base config fix on Windows and minor changes in setup.py
@@ -14,10 +26,9 @@
   - Added proper version of lxml in requirements
 
 ## 0.2.1
-  - Added pass/fail criteria report on console reporter.
+  - Added pass/fail criteria report on console reporter
 
 ## 0.2.0
-
   - Added JSON path assertion.
   - Added parameters for final_stats reporter
   - Added ability to generate query string based on parameters and request type.